% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar_datos_icon.R
\name{cargar_datos_icon}
\alias{cargar_datos_icon}
\title{Cargar y Procesar Datos del Modelo ICON}
\usage{
cargar_datos_icon(
  variable,
  fecha_inicio,
  fecha_fin,
  ruta,
  utc_offset = 4,
  hora_modelo = "00",
  acum = FALSE,
  ExtChile = FALSE
)
}
\arguments{
\item{variable}{Character. Nombre de la variable a procesar. Las opciones incluyen "TOT_PREC", "snow_gsp", "snow_con", "h_snow", "rain_con", "rain_gsp" para datos de precipitación, y "T_2M", "TMIN_2M", "TMAX_2M" para datos de temperatura.}

\item{fecha_inicio}{Date. Fecha de inicio del rango de datos.}

\item{fecha_fin}{Date. Fecha de fin del rango de datos.}

\item{ruta}{Character. Ruta del directorio donde se encuentran los archivos .nc.}

\item{utc_offset}{Numeric. Desplazamiento UTC en horas para ajustar los tiempos de los datos.}

\item{hora_modelo}{Character. Hora del modelo, usualmente "00".}

\item{acum}{Logical. Indica si se debe realizar un acumulado diario para las variables de precipitación. El valor predeterminado es FALSE.}

\item{ExtChile}{Logical. Indica si se debe recortar los datos a la extensión de Chile. El valor predeterminado es FALSE.}
}
\value{
Un objeto SpatRaster con los datos procesados. Si \code{acum} es TRUE, se devuelve el acumulado diario.
}
\description{
Esta función carga y procesa datos del modelo ICON según la variable especificada. La función maneja tanto datos de temperatura como de precipitación, y ajusta los tiempos de acuerdo con un desplazamiento UTC especificado.
}
\examples{
ruta <- "E:/ICON_INFORME/ICON"
fecha_inicio <- as.Date("2024-06-05")
fecha_fin <- as.Date("2024-06-07")
variable <- "TOT_PREC"
hora_modelo <- "00"

datos <- cargar_datos_icon(variable, fecha_inicio, fecha_fin, ruta, utc_offset = 4, hora_modelo = hora_modelo, acum = TRUE, ExtChile = TRUE)

print(datos)

nuble <- cargar_regiones("R16")
nuble <- terra::project(nuble, datos)
icon_daily_nuble <- terra::crop(datos[[1]], nuble, mask = TRUE, touches = TRUE)

plot(icon_daily_nuble)

}
