% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar_regiones.R
\name{cargar_regiones}
\alias{cargar_regiones}
\title{Cargar Regiones Simplificadas}
\usage{
cargar_regiones(regiones = NULL)
}
\arguments{
\item{regiones}{Vector de caracteres opcional que especifica las regiones que se desean cargar.}
}
\value{
Un objeto \code{SpatVector} si se carga una sola región o todas las regiones unidas, o una lista de objetos \code{SpatVector} si se cargan múltiples regiones.
}
\description{
Esta función permite cargar shapefiles simplificados de regiones específicas de Chile almacenados dentro del paquete \code{iconDWD}.
}
