% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertirRutaWindowsAWSL.R
\name{convertirRutaWindowsAWSL}
\alias{convertirRutaWindowsAWSL}
\title{Convertir Ruta de Windows a WSL}
\usage{
convertirRutaWindowsAWSL(rutaWindows)
}
\arguments{
\item{rutaWindows}{Una cadena de texto que representa la ruta de archivo en el formato de Windows, como "C:/Users/NombreUsuario/Documentos".}
}
\value{
Una cadena de texto que representa la ruta de archivo convertida en el formato de WSL, como "/mnt/c/Users/NombreUsuario/Documentos".
}
\description{
Esta función toma una ruta de archivo de Windows y la convierte en una ruta compatible con el Subsistema de Windows para Linux (WSL).
}
\details{
La función primero normaliza las barras de la ruta de Windows a barras hacia adelante.
Luego, extrae la letra de la unidad y el resto de la ruta para construir la nueva ruta en el formato de WSL.
Es importante notar que esta función asume que se está trabajando con el Subsistema de Windows para Linux y
que las rutas de unidades en WSL se montan bajo "/mnt/" seguido de la letra de la unidad en minúsculas.
}
\examples{
convertirRutaWindowsAWSL("C:/Users/NombreUsuario/Documentos")
}
