% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_valores_icon.R
\name{extraer_valores_icon}
\alias{extraer_valores_icon}
\title{Extraer Valores y Fechas de un SpatRaster para Ubicaciones Específicas}
\usage{
extraer_valores_icon(spatraster, coords, ...)
}
\arguments{
\item{spatraster}{El objeto SpatRaster del cual se extraerán los valores.}

\item{coords}{Coordenadas de la localización para la extracción, que pueden ser un punto (matrix o data.frame con
columnas x, y) o un SpatVector (puntos, líneas, polígonos).}

\item{...}{Argumentos adicionales para la función \code{extract}.}

\item{fun}{(Opcional) Función para resumir los datos extraídos por línea o geometría de polígono.}

\item{touches}{(Opcional) Lógico, si TRUE, se extraen valores para todas las celdas tocadas por líneas o polígonos.}
}
\value{
Un dataframe con dos columnas: Fecha y Valor, representando las fechas de cada capa
del SpatRaster y los valores extraídos para la localización dada.
}
\description{
Esta función extrae los valores de un SpatRaster para una localización dada (punto o SpatVector)
y los formatea en un dataframe con las correspondientes fechas y valores. Soporta argumentos adicionales
de la función \code{extract} de \code{terra}, como \code{fun}, \code{touches}, entre otros.
}
\examples{
coords_estacion <- obtener_coords_estacion(nombre_estacion = "Ce arroz")
valores_y_fechas <- extraer_valores_y_fechas(icon_18, coords_estacion)
print(valores_y_fechas)

}
