% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraerBZ2.R
\name{extraerBZ2}
\alias{extraerBZ2}
\title{Extraer archivos .bz2 en paralelo o secuencialmente}
\usage{
extraerBZ2(rutas, parallel = FALSE, ncores = parallel::detectCores() - 1)
}
\arguments{
\item{rutas}{Vector de rutas completas a los archivos .bz2 que se desean extraer.}

\item{parallel}{Lógico, indica si se debe ejecutar en paralelo.}

\item{ncores}{Número de núcleos a utilizar para la ejecución paralela.}
}
\value{
Imprime un mensaje indicando el resultado de la extracción.
}
\description{
Esta función verifica si un archivo o archivos .bz2 existen en la ruta especificada y, en caso afirmativo,
los extrae en el mismo directorio de manera opcionalmente paralela.
}
\details{
Si \code{parallel} es TRUE, la función configura un clúster usando el paquete doParallel
y extrae los archivos en paralelo utilizando el número especificado de núcleos.
Requiere que \code{bunzip2} esté instalado y disponible en el sistema.
}
\examples{
extraerBZ2(c("path/to/your/file1.bz2", "path/to/your/file2.bz2"), parallel = TRUE, ncores = 2)
}
