% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiltrarVariable.R
\name{FiltrarVariable}
\alias{FiltrarVariable}
\title{Filtra archivos por variable y rango de fechas}
\usage{
FiltrarVariable(dir, fecha_inicial = NA, fecha_final = NA, variable)
}
\arguments{
\item{dir}{Directorio que contiene los archivos a filtrar.}

\item{fecha_inicial}{Fecha de inicio del rango a filtrar en formato "YYYY-MM-DD HH:MM:SS" (opcional).
Si se omite, no se aplica filtro de fecha de inicio.}

\item{fecha_final}{Fecha final del rango a filtrar en formato "YYYY-MM-DD HH:MM:SS" (opcional).
Si se omite, no se aplica filtro de fecha final.}

\item{variable}{Nombre de la variable en mayúsculas por la cual filtrar, como TMIN_2M.}
}
\value{
Un vector con las rutas completas de los archivos que cumplen con los criterios.
}
\description{
Esta función filtra archivos de proyecciones climáticas por una variable específica y un rango de fechas opcional.
Las variables admitidas en mayúsculas son: TMIN_2M, TMAX_2M, TOT_PREC, WW, CAPE_ML, T_2M.
Devuelve las rutas completas de los archivos que cumplen con los criterios especificados.
}
\examples{
\dontrun{
FiltrarVariable(dir = "path/to/your/files",
                fecha_inicial = "2024-03-19 00:00:00",
                fecha_final = "2024-03-20 00:00:00",
                variable = "T_2M")
}
}
