% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grib2ANetCDF.R
\name{Grib2ANetCDF}
\alias{Grib2ANetCDF}
\title{Convertir archivos .grib2 a .nc}
\usage{
Grib2ANetCDF(
  ruta_in,
  ruta_out,
  ruta_script,
  parallel = FALSE,
  ncores = detectCores() - 3,
  verbose = TRUE
)
}
\arguments{
\item{ruta_in}{Vector de caracteres con las rutas de los archivos de entrada .grib2.}

\item{ruta_out}{Vector de caracteres con las rutas de destino para los archivos .nc.}

\item{ruta_script}{Ruta del script dentro de wsl("/home/...")}

\item{parallel}{Lógico, indica si la conversión debe realizarse en paralelo. Por defecto es \code{FALSE}.}

\item{ncores}{Entero, número de núcleos a utilizar para la conversión en paralelo. Solo aplica si \code{parallel} es \code{TRUE}. Por defecto, usa todos los núcleos disponibles menos uno.}

\item{verbose}{Lógico, indica si la función debe imprimir mensajes sobre el progreso de la conversión. Por defecto es \code{TRUE}.}
}
\value{
La función no retorna un objeto. Los archivos .nc son escritos en las rutas especificadas
por \code{ruta_out}.
}
\description{
Esta función convierte archivos de formato .grib2 a formato .nc (NetCDF) utilizando
una llamada al sistema para ejecutar un script de bash en WSL. La función puede
operar en modo secuencial o paralelo.
}
\details{
Para el funcionamiento en modo paralelo, esta función depende del paquete \code{doSNOW} y
configura un clúster SOCK para la ejecución paralela de las conversiones. La conversión
de rutas de Windows a WSL se realiza mediante la función \code{convertirRutaWindowsAWSL},
que debe estar definida previamente o ser parte del mismo paquete.

La barra de progreso y el cálculo del tiempo estimado se añaden para mejorar la
interacción del usuario con la función, especialmente útil para conversiones de
larga duración.
}
\examples{
\dontrun{
ruta_in <- c("/path/to/input/file1.grib2", "/path/to/input/file2.grib2")
ruta_out <- c("/path/to/output/file1.nc", "/path/to/output/file2.nc")
ruta_script_wsl <- "/home/inia/ICON_0125/transform_0125.sh"
Grib2ANetCDF(ruta_in, ruta_out, ruta_script, parallel = TRUE, verbose = TRUE)
}

}
