% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_coords_estacion.R
\name{obtener_coords_estacion}
\alias{obtener_coords_estacion}
\title{Obtener Coordenadas de Estación Meteorológica}
\usage{
obtener_coords_estacion(nombre_estacion)
}
\arguments{
\item{nombre_estacion}{Cadena de caracteres con el nombre de la estación meteorológica.}
}
\description{
Busca las coordenadas de una estación meteorológica en un archivo CSV incluido en el paquete,
permitiendo una búsqueda flexible del nombre de la estación.
}
